requirements.txt
openai==1.40.6
langchain==0.2.14
langchain-openai==0.1.21
python-dotenv==1.0.1
pymssql==2.2.7